\name{adci}
\alias{adci}
\title{
Abundance confidence intervals}
\description{
Derive the abundance confidence intervals over the simulated arena for a stage and a certain date or range of dates.}
\usage{
adci(input_sim=NA, stage=1, eval_date=0, breaks=c(0.25,0.5,0.75),np=1)
}
\arguments{
\item{input_sim}{input data.}
\item{stage}{input data.}
\item{eval_date}{input data.}
\item{breaks}{input data.}
\item{np}{input data.}
}
\value{
To be done
}
\author{
Daniele Da Re \email{daniele.dare@uclouvain.be} \cr
Marcantonio Matteo \email{marcantoniomatteo@gmail.com} \cr
}
\keyword{internal}
