\name{adci_sp}
\alias{adci}
\title{
Abundance confidence intervals}
\description{
Derive abundance confidence intervals for each pixel in the simulated arena for a stage and a certain date or range of dates.}
\usage{
dynamAedes(coords=NULL, input_sim=NULL, eval.date=NULL, st=1,  breaks=NULL)
}
\arguments{
\item{coords}{input data.}
\item{input_sim}{input data.}
\item{eval_date}{input data.}
\item{st}{input data.}
\item{breaks}{input data.}
}
\value{A RasterStack or a list of RasterStack.
}
\author{
Daniele Da Re \email{daniele.dare@uclouvain.be} \cr
Marcantonio Matteo \email{marcantoniomatteo@gmail.com} \cr
}
\keyword{internal}
