% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AedeslifeHistoryList.R
\docType{data}
\name{AedeslifeHistoryList}
\alias{AedeslifeHistoryList}
\title{Reference list for \strong{Aedes} stages and substages}
\format{
A list of two matrices
}
\usage{
data(AedeslifeHistoryList)
}
\description{
Reference files to match simulation output matrices with
\strong{Aedes} stages and substages as modelled in dynamAedes.
The list contains:
}
\details{
\itemize{
\item speciesheet: description of each substage for each species
\item codesheet: reference coding to match substages to simulation output
}
}
\keyword{cheat}
\keyword{list}
