% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{check_date_format}
\alias{check_date_format}
\title{Date format check}
\usage{
check_date_format(date)
}
\value{
Logical
}
\description{
Check the proper format of a date object
}
\keyword{internal}
