% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dici.R
\name{dici}
\alias{dici}
\title{Estimate of of mosquito dispersal spread}
\usage{
dici(
  input_sim = NULL,
  coords = NULL,
  eval_date = NULL,
  breaks = c(0.25, 0.5, 0.75),
  space = FALSE
)
}
\arguments{
\item{input_sim}{(matrix) dynamAedes compressed output matrix}

\item{coords}{(matrix) (matrix) a matrix reporting the spatial coordinates of the temperature observations.}

\item{eval_date}{(integer) define the day of successful introduction evaluation, referring to the column number of the temperature matrix used to inform the model.}

\item{breaks}{quantile breaks, default the first, the second and the third c(0.25,0.5,0.75)}

\item{space}{is it spatial?}
}
\value{
TBW
}
\description{
Estimates of mosquito dispersal spread (in km2 ) of the simulated mosquito populations when scale = "lc"
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
