% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.a.a_disp.f}
\alias{.a.a_disp.f}
\title{Log-Normal probability density of short active dispersal}
\usage{
.a.a_disp.f(sp, max.a.disp, disp.bins)
}
\value{
vector of rates.
}
\description{
Log-Normal probability density of short active dispersal from Marcantonio et al. 2019; Marini et al. 2019. 0 to 600 m with resolution of 10 m.
}
\keyword{internal}
