% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.a.gono_rate.f}
\alias{.a.gono_rate.f}
\title{Daily rate of gonotrophic cycle}
\usage{
.a.gono_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Daily rate of gonotrophic cycle, i.e., daily rate of adult females completing the cycle from bloodmeal to the first day of oviposition
}
\keyword{internal}
