% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.a.ovi_rate.f}
\alias{.a.ovi_rate.f}
\title{Oviposition rate}
\usage{
.a.ovi_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Oviposition rate, i.e., number of eggs laid per female/day at different temperature
}
\keyword{internal}
