% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.a.surv_rate.f}
\alias{.a.surv_rate.f}
\title{Adult female daily mortality rate}
\usage{
.a.surv_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Adult female daily mortality rate at different temperatures
}
\keyword{internal}
