% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.beta.fun}
\alias{.beta.fun}
\title{Function 1 to start Beta functions: Functions taken from https://github.com/OnofriAndreaPG/aomisc}
\usage{
.beta.fun(X, b, d, Xb, Xo, Xc)
}
\value{
No return value, called for side effects in .DRC.beta
}
\description{
Function 1 to start Beta functions
}
\keyword{internal}
