% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.beta.init}
\alias{.beta.init}
\title{Function 3 to start Beta functions: Functions taken from https://github.com/OnofriAndreaPG/aomisc}
\usage{
.beta.init(mCall, LHS, data)
}
\value{
No return value, called for Beta function initiation in drm
}
\description{
Function 1 to start Beta functions
}
\keyword{internal}
