% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.d.surv_rate.f}
\alias{.d.surv_rate.f}
\title{Diapause eggs daily survival rate}
\usage{
.d.surv_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Diapause eggs daily survival rate at different temperature
}
\keyword{internal}
