% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.DRC.beta}
\alias{.DRC.beta}
\title{Function 2 to start Beta functions: Functions taken from https://github.com/OnofriAndreaPG/aomisc}
\usage{
.DRC.beta()
}
\value{
No return value, called for side effects in drm
}
\description{
Function 1 to start Beta functions
}
\keyword{internal}
