% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.e.dia_rate.f}
\alias{.e.dia_rate.f}
\title{Allocation of diapause eggs}
\usage{
.e.dia_rate.f(photo.new, sp)
}
\value{
vector of rates.
}
\description{
Allocation of diapause eggs dependent on photoperiod
}
\keyword{internal}
