% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.e.hatch_rate.f}
\alias{.e.hatch_rate.f}
\title{Egg hatching rate}
\usage{
.e.hatch_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Egg hatching rate: this rate decides embryonated eggs which hatch or stay
}
\keyword{internal}
