% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.e.surv_rate.f}
\alias{.e.surv_rate.f}
\title{Egg daily survival rate}
\usage{
.e.surv_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Egg daily survival rate at different temperature
}
\keyword{internal}
