% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.i.ddmort_rate.f}
\alias{.i.ddmort_rate.f}
\title{Immature Density-dependent mortality rate}
\usage{
.i.ddmort_rate.f(dens.new)
}
\value{
vector of rates.
}
\description{
Immature Density-dependent mortality
}
\keyword{internal}
