% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.i.emer_rate.f}
\alias{.i.emer_rate.f}
\title{Immature daily emergence rate}
\usage{
.i.emer_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Immature daily emergence rate at different temperature
}
\keyword{internal}
