% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.i.surv_rate.f}
\alias{.i.surv_rate.f}
\title{Immature daily survival rat}
\usage{
.i.surv_rate.f(temp.new, sp)
}
\value{
vector of rates.
}
\description{
Immature daily survival rate at different temperature
}
\keyword{internal}
