% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.meuc}
\alias{.meuc}
\title{Euclidean distance from a pair of coordinates}
\usage{
.meuc(c1, c2, coords)
}
\value{
float. Vector of Euclidean distances.
}
\description{
Returns Euclidean distance from a pair of coordinates
}
\keyword{internal}
