% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.resample}
\alias{.resample}
\title{Safer version of sample}
\usage{
.resample(x, ...)
}
\value{
a vector of length n with elements drawn from either x or from the integers 1:x
}
\description{
sample
}
\keyword{internal}
