% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-helpers.R
\name{.returndis}
\alias{.returndis}
\title{Distance moved by mosquito populations}
\usage{
.returndis(distl = NA, days = NA, breaks = breaks)
}
\value{
vector of quantiles of distances.
}
\description{
Return quantiles of distance moved by mosquito populations
}
\keyword{internal}
