% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamAedesClass.R
\docType{class}
\name{dynamAedesClass-class}
\alias{dynamAedesClass-class}
\title{S4 class representing the output of dynamAedes.m}
\description{
S4 class representing the output of dynamAedes.m
}
\section{Slots}{

\describe{
\item{\code{species}}{Character. The simulated species.}

\item{\code{scale}}{Character. The scale of the simulation.}

\item{\code{start_date}}{Character. The introduction date.}

\item{\code{end_date}}{Character. The end date of the simulation.}

\item{\code{n_iterations}}{Numeric. The number of iterations.}

\item{\code{stage_intro}}{Character. The introduced mosquito stage.}

\item{\code{n_intro}}{numeric. The number of propagules introduced.}

\item{\code{coordinates}}{Matrix. The coordinates of each cell.}

\item{\code{compressed_output}}{Logical. If output is at stage or substage level.}

\item{\code{jhwv}}{Numeric. The volume of water in the system.}
}}

