% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates_spatial.R
\name{get_rates_spatial}
\alias{get_rates_spatial}
\title{Proportion of successful introductions}
\usage{
get_rates_spatial(
  coords = NULL,
  daily_temp = NULL,
  species = NULL,
  rate_fun = NULL,
  spatial = FALSE,
  rate = TRUE,
  np = 1
)
}
\arguments{
\item{coords}{bla}

\item{daily_temp}{bla}

\item{species}{bla}

\item{rate_fun}{bla}

\item{spatial}{bla}

\item{rate}{bla}

\item{np}{bla}
}
\value{
bla
}
\description{
Compute the proportion of "successful" introductions.
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
