% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icci.R
\name{icci}
\alias{icci}
\title{Number of invaded cells}
\usage{
icci(input_sim = NA, eval_date = 0, breaks = c(0.25, 0.5, 0.75))
}
\arguments{
\item{input_sim}{matrix. \code{dynamAedes} compressed output matrix (\code{compressed=TRUE}).}

\item{eval_date}{positive integer. Define the day of successful introduction evaluation, referring to the column number of the temperature matrix used to inform the model.}

\item{breaks}{numeric vector. Quantile breaks, default the first, the second and the third quantile: \code{c(0.25,0.5,0.75)}.}
}
\value{
\code{icci} returns the Interquartile range of the number of invaded cells for the specified day over model all iterations. The output should be interpreted according to dynamAedes spatial scale (i.e. \code{scale='rg'} or  \code{scale='lc'}).
}
\description{
Compute a summary of the number of invaded cells over model iterations
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
