% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamAedesClass.R
\name{max,dynamAedesClass-method}
\alias{max,dynamAedesClass-method}
\alias{max,dynamAedesClass-max}
\title{Max method for dynamAedesClass
Provides the max number of days with at least one
propagule in the system (any stage) along iterations.}
\usage{
\S4method{max}{dynamAedesClass}(x, na.rm = FALSE)
}
\arguments{
\item{x}{An object of class \code{dynamAedesClass}.}

\item{na.rm}{logic.}
}
\value{
An integer.
}
\description{
Max method for dynamAedesClass
Provides the max number of days with at least one
propagule in the system (any stage) along iterations.
}
\examples{
\dontrun{
max(sim)
}

}
