% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamAedesClass.R
\name{min,dynamAedesClass-method}
\alias{min,dynamAedesClass-method}
\alias{min,dynamAedesClass-min}
\title{Min method for dynamAedesClass
Provides the min number of days with at least one
propagule in the system (any stage) along iterations.}
\usage{
\S4method{min}{dynamAedesClass}(x, na.rm = FALSE)
}
\arguments{
\item{x}{An object of class \code{dynamAedesClass}.}

\item{na.rm}{logic.}
}
\value{
An integer.
}
\description{
Min method for dynamAedesClass
Provides the min number of days with at least one
propagule in the system (any stage) along iterations.
}
\examples{
\dontrun{
min(sim)
}

}
