% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleIntro.R
\name{multipleIntro}
\alias{multipleIntro}
\title{Multiple introductions of Aedes sp. propagules}
\usage{
multipleIntro(
  startd = NULL,
  endd = NULL,
  n.intro = NULL,
  intro.eggs = NULL,
  intro.adults = NULL,
  sd_eggs = NULL,
  scenario = "ff",
  iter = NULL
)
}
\arguments{
\item{startd}{character. The date of the start of simulations, also corresponding to the first introduction event of mosquito propagules.}

\item{endd}{character. The date of the end of simulations.}

\item{n.intro}{numeric. The number of times propagules will be introduced throughout the simulations.}

\item{intro.eggs}{numeric. The average number of eggs introduced per single introduction event. Note: specify either \code{intro.eggs} or \code{intro.adults}. If both are specified, \code{intro.eggs} will be used as default.}

\item{intro.adults}{numeric. The average number of female mosquitoes introduced per single introduction event. Note: specify either \code{intro.eggs} or \code{intro.adults}. If both are specified, \code{intro.eggs} will be used as default.}

\item{sd_eggs}{numeric. The standard deviation around the average number of eggs introduced by \code{intro.eggs} or produced by \code{intro.adults} per single introduction event.}

\item{scenario}{character. The number of eggs and frequency of their introduction scenario. The letter word defines the frequency of egg introduction, which is either at "fixed" or "random" intervals, thus "f" of "r". The second word defines the number of eggs introduced, which is either "fixed" or "random",  thus "f" of "r". The latter corresponds to a number randomly sampled from a normal distribution with a mean defined by either \code{intro.eggs} or \code{intro.adults}. Default \code{scenario = "ff"}.}

\item{iter}{positive integer. The number of model iterations.}
}
\value{
A list of vectors containing the number of eggs introduced each time and the date at which said eggs are introduced. The length of the list is equal to the number of iterations.
}
\description{
\code{multipleIntro} allows testing scenarios of multiple introduction of invasive mosquitoes.
}
\details{
For a given iteration, the function returns the number of eggs introduced and the dates at which they are introduced.
The function allows to introduce mosquito eggs multiple times for each model iteration, determining how many eggs are introduced and when, based on four possible scenarios:
\itemize{
\item \code{"ff"}, i.e. "fixed fixed": egg introductions at fixed intervals in time, and fixed number of eggs introduced each time
\item \code{"fr"}, i.e. "fixed random": egg introductions at fixed intervals in time, but random number of eggs introduced each time
\item \code{"rf"}, i.e. "random fixed": egg introductions at random intervals in time, but fixed number of eggs introduced each time
\item \code{"rr"}, i.e. "random random": egg introductions at random intervals in time, and random number of eggs introduced each time
}

The function allows introducing two different mosquito life history stages: egg and female adult. By introducing adults, the number of eggs produced by a single female is sampled from a normal distribution
with \eqn{\mu = 77} and \eqn{\sigma = 15}, being based on the findings presented in Aida et al. (2011)\url{https://doi.org/10.1016/S2221-1691(11)60103-2}
The \code{"fr"} and \code{"rr"} scenarios currently support only the species Aedes albopictus (Aida et al.,2011).
}
\examples{
\dontrun{out <- multipleIntro(startd = "2022-06-10",
                       endd = "2022-08-10",
                       n.intro = 4,
                       intro.eggs = 100,
                       sd_eggs = 15,
                       scenario = "ff",
                       iter = 100)
}
}
\author{
Emmanuelle Kern \email{emmanuelle.kern22@imperial.ac.uk}, Ilaria Dorigatti  \email{i.dorigatti@imperial.ac.uk}
}
