% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamAedesClass.R
\name{summary,dynamAedesClass-method}
\alias{summary,dynamAedesClass-method}
\alias{summary,dynamAedesClass-summary}
\title{Summary method for dynamAedesClass
Provides a summary of simulations based on the dynamAedesClass.}
\usage{
\S4method{summary}{dynamAedesClass}(object)
}
\arguments{
\item{object}{An object of class \code{dynamAedesClass}.}
}
\value{
A character vector with the summary details of the simulation.
}
\description{
Summary method for dynamAedesClass
Provides a summary of simulations based on the dynamAedesClass.
}
\examples{
\dontrun{
summary(sim)
}

}
