\name{adci}
\alias{adci}
\title{
Dispersal confidence intervals per day}
\description{
Derive the dispersal distance confidence intervals from introduction cell for a certain date or range of dates.}
\usage{
adci(input_sim=NA, coords=NULL, eval_date=0, breaks=c(0.25,0.5,0.75),space=FALSE)
}
\arguments{
\item{input_sim}{input data.}
\item{coords}{input data.}
\item{eval_date}{input data.}
\item{breaks}{input data.}
\item{space}{input data.}
}
\value{
To be done
}
\author{
Marcantonio Matteo \email{marcantoniomatteo@gmail.com} \cr
Daniele Da Re \email{daniele.dare@uclouvain.be} \cr
}
\keyword{internal}
