% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamAedes.R
\name{dynamAedes}
\alias{dynamAedes}
\title{Life cycle simulation of \emph{Aedes} mosquitoes}
\usage{
dynamAedes(
  species = "aegypti",
  intro.eggs = 0,
  intro.adults = 0,
  intro.juveniles = 0,
  scale = "ws",
  intro.cells = NULL,
  ihwv = 1,
  temps.matrix = NULL,
  startd = 1,
  endd = 10,
  cells.coords = NULL,
  lat = 0,
  long = 0,
  road.dist.matrix = NULL,
  country = NA,
  intro.year = 2020,
  iter = 1,
  n.clusters = 1,
  cluster.type = "PSOCK",
  sparse.output = FALSE,
  compressed.output = TRUE,
  suffix = NA,
  cellsize = 250,
  maxadisp = 600,
  dispbins = 10,
  verbose = FALSE
)
}
\arguments{
\item{species}{select what species to model: "aegypti", "albopictus", "japonicus", "koreicus". Default species = "aegypti".}

\item{intro.eggs}{(numeric) number of introduced propagules, default intro.eggs = 100.}

\item{intro.adults}{(numeric) number of introduced propagules, default zero.}

\item{intro.juveniles}{(numeric) number of introduced propagules, default zero.}

\item{scale}{define the spatial scale of the modelling exercise: punctual/weather station "ws", local "lc", or regional "rg". Active dispersal is enabled only for scale = "rg". Default scale = "ws".}

\item{intro.cells}{bla.}

\item{ihwv}{(numeric) larval-habitat water volume, define the volume (L) of water habitat presents in each spatial unit (parameterised from \href{https://doi.org/10.1111/1365-2664.12620}{Hancock et al. 2016}). Default lhwv = 1.}

\item{temps.matrix}{(matrix) a matrix of daily average temperature (in Celsius degree /1000) used to fit the life cycle rates. This matrix must be organised with the daily temperature observations as columns and the geographic position of the i-grid cell as rows.}

\item{startd}{(numeric) day of start of the simulations, referring to the column number of temps.matrix.}

\item{endd}{(numeric) day of end of the simulation, referring to the column number of temps.matrix.}

\item{cells.coords}{(matrix) a matrix reporting the spatial coordinates of the temperature observations.}

\item{lat}{(numeric) latitude value of the area of interested used to inform the diapause eggs allocation function.}

\item{long}{(numeric) longitude value of the area of interested used to inform the diapause eggs allocation function.}

\item{road.dist.matrix}{(matrix) when scale = "lc", define the matrix containing the coordinates of the grid cells of the landscape intersecting the road network for the mosquito passive dispersal process.}

\item{country}{when scale = "lc", define the average car trip distance for the mosquito passive dispersal process. The value can be provided by the users, or using the estimates made by \href{https://publications.jrc.ec.europa.eu/repository/handle/JRC77079}{Pasaoglu et al. 2012}). for the following Europeancountries: France "fra", Germany "deu", Italy "ita", Poland "pol", Spain "esp", and the United Kingdom "uk".}

\item{intro.year}{(numeric) year of the beginning of iterations.}

\item{iter}{(numeric) define the number of model iteration.}

\item{n.clusters}{(numeric) define the number of parallel processes.}

\item{cluster.type}{define the type of cluster, default "PSOCK".}

\item{sparse.output}{BHOOOOOOOOOOOO}

\item{compressed.output}{(logical) default TRUE, if FALSE provide the individual estimates for each iterations and each model's subcompartiment.}

\item{suffix}{(character) model output suffix.}

\item{cellsize}{(numeric) when scale = "lc", define the minimal distance of the active dispersal kernel and should match the spatial resolution of temps.matrix to avoid inconsistencies. Default cellsize = 250}

\item{maxadisp}{(numeric) when scale = "lc", define the maximum daily dispersal, default maxadisp = 600.}

\item{dispbins}{(numeric) when scale = "lc", define the resoluzion of the dispersal kernel, default dispbins = 10.}

\item{verbose}{(logical)}
}
\value{
Matrix or a list of matrixes containing, for each iteration, the number of individuals in each life stage per day (and for each grid cell of the study area if scale="lc" or "rg"). If the argument compressed.output=FALSE (default TRUE), the model returns the daily number of individuals in each life stage sub-compartment.
}
\description{
Function to simulate population dynamics of \emph{Aedes} mosquitoes
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
