\name{get_rates_spatial}
\alias{get_rates_spatial}
\title{
Spatial physiological temperature dependent rates}
\description{
Get physiological temperature dependent rates per day in each pixel.}
\usage{
get_rates_spatial(coords=NULL, daily_temp=NULL, species=NULL, rate_fun=NULL, spatial=FALSE, rate=TRUE)
}
\arguments{
\item{coords}{input data.}
\item{daily_temp}{input data.}
\item{species}{input data.}
\item{rate_fun}{input data.}
\item{spatial}{input data.}
\item{rate}{input data.}
}
\value{
To be done
}
\author{
Daniele Da Re \email{daniele.dare@uclouvain.be} \cr
Marcantonio Matteo \email{marcantoniomatteo@gmail.com} \cr
}
\keyword{internal}
