\name{icci}
\alias{icci}
\title{
Confidence intervals of number invaded cells}
\description{
Confidence intervals of number invaded cells}
\usage{
dynamAedes(input_sim=NA, days=0, breaks=c(0.25,0.5,0.75))
}
\arguments{
\item{input_sim}{input data.}
\item{days}{input data.}
\item{breaks}{input data.}
}
\value{
To be done
}
\author{
Daniele Da Re \email{daniele.dare@uclouvain.be} \cr
Marcantonio Matteo \email{marcantoniomatteo@gmail.com} \cr
}
\keyword{internal}
