% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icci.R
\name{icci}
\alias{icci}
\title{Number of invaded cells}
\usage{
icci(input_sim = NA, eval_date = NULL, breaks = c(0.25, 0.5, 0.75))
}
\arguments{
\item{input_sim}{matrix. \code{dynamAedes.m} compressed output matrix (\code{compressed=TRUE}).}

\item{eval_date}{numeric. Define the day of evaluation; it refers to the column number of the input temperature matrix.}

\item{breaks}{numeric vector. Quantile breaks, default interquartile range: \code{c(0.25,0.5,0.75)}.}
}
\value{
\code{icci} returns quantiles of the distribution of the invaded cell number for the specified. The output should be interpreted according to model spatial scale (i.e. \code{scale='rg'} or \code{scale='lc'} give different interpretation).
}
\description{
Compute a summary of the number of invaded cells over model iterations
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
