% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_sp.R
\name{psi_sp}
\alias{psi_sp}
\title{Probability of successful introduction (spatial)}
\usage{
psi_sp(input_sim = NULL, eval_date = NULL, n.clusters = 1)
}
\arguments{
\item{input_sim}{matrix. \code{dynamAedes.m} compressed output matrix (\code{compressed=TRUE}).}

\item{eval_date}{positive integer. Define the day(s) to calculate the proportion of successful introductions which should match the column number of the temperature matrix used to inform the model.}

\item{n.clusters}{positive integer. Define the number of parallel processes.}
}
\value{
\code{psi_sp} returns a raster with the proportion of model iterations that resulted in a viable mosquito population at a given date for a given life stage in each cell of the grid.
}
\description{
Compute the proportion of successful introductions per each cell of the grid.
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, Daniele Da Re \email{daniele.dare@uclouvain.be}
}
